use data offload panel instead of download panel
